chore(cli): fixes and features

+++ crates/cli/src/config/processor/update.rs
@@ -8,6 +8,7 @@ use crate::config::models::Config as ConfigModel;
 use crate::config::models::Profile;
 use crate::config::processor::prompt::profile_prompt;
 use crate::config::{get_profile, read_config, write_config, ConfigError};
+use crate::prompter;
 
 pub fn process_update(
     profile_name: String,
@@ -18,6 +19,14 @@ pub fn process_update(
     let path = Path::new(&config_path);
 
     if path.exists() && reset {
+        let confirm = prompter::confirm(
+            "Existing config will be deleted. Do you want to proceed?",
+            Some(false),
+        )?;
+        if !confirm {
+            println!("Aborting configuration. Use `config update` command to update the config.");
+            return Ok(());
+        }
         std::fs::remove_file(path)?;
     }
